libxtreemfs: reverted 3156
git-svn-id: https://xtreemfs.googlecode.com/svn/trunk@3157 f8528748-c2c0-11dd-ba5c-d3b7c2b3b534

+++ cpp/src/fuse/fuse_adapter.cpp
@@ -790,7 +790,7 @@ int FuseAdapter::releasedir(const char *path, struct fuse_file_info *fi) {
   assert(cached_direntries != NULL);
   delete cached_direntries->dir_entries;
   delete cached_direntries;
-  fi->fh = static_cast<uint64_t>(NULL);
+  fi->fh = NULL;
 
   return 0;
 }
@@ -1254,19 +1254,19 @@ int FuseAdapter::release(const char *path, struct fuse_file_info *fi) {
     try {
       file_handle->Close();
     } catch(const PosixErrorException& e) {
-      fi->fh = static_cast<uint64_t>(NULL);
+      fi->fh = NULL;
       return -1 * ConvertXtreemFSErrnoToFuse(e.posix_errno());
     } catch(const XtreemFSException& e) {
-      fi->fh = static_cast<uint64_t>(NULL);
+      fi->fh = NULL;
       return -1 * EIO;
     } catch(const exception& e) {
       ErrorLog::error_log->AppendError("A non-XtreemFS exception occured: "
                 + string(e.what()));
-      fi->fh = static_cast<uint64_t>(NULL);
+      fi->fh = NULL;
       return -1 * EIO;
     }
   }
-  fi->fh = static_cast<uint64_t>(NULL);
+  fi->fh = NULL;
   return 0;
 }
